Adjustments for confidence calculations
Change the calculation of the confidence weights so that the weight for a
normalized vector becomes one. The idea behind this is so that enabling
--confidence by itself has no impact on the final result, the final
surface is only changed if there are normals with additional weight.

Implementation wise, previously the normal vector was transformed causing
the length of the vector to be changed. This change calculates the
original length, applies the transformation, and then changes the length
to the original one.